Struct unix_socket::UnixDatagram

pub struct UnixDatagram {
    // some fields omitted
}

A Unix datagram socket.

Examples

use unix_socket::UnixDatagram;

let socket = UnixDatagram::bind("/path/to/my/socket").unwrap();
socket.send_to(b"hello world", "/path/to/other/socket").unwrap();
let mut buf = [0; 100];
let (count, address) = socket.recv_from(&mut buf).unwrap();
println!("socket {:?} sent {:?}", address, &buf[..count]);

Methods

impl UnixDatagram

fn bind<P: AsRef<Path>>(path: P) -> Result<UnixDatagram>

Creates a Unix datagram socket bound to the given path.

Linux provides, as a nonportable extension, a separate "abstract" address namespace as opposed to filesystem-based addressing. If path begins with a null byte, it will be interpreted as an "abstract" address. Otherwise, it will be interpreted as a "pathname" address, corresponding to a path on the filesystem.

fn unbound() -> Result<UnixDatagram>

Creates a Unix Datagram socket which is not bound to any address.

fn pair() -> Result<(UnixDatagram, UnixDatagram)>

Create an unnamed pair of connected sockets.

Returns two UnixDatagramss which are connected to each other.

fn connect<P: AsRef<Path>>(&self, path: P) -> Result<()>

Connect the socket to the specified address.

The send method may be used to send data to the specified address. recv and recv_from will only receive data from that address.

fn local_addr(&self) -> Result<SocketAddr>

Returns the address of this socket.

fn peer_addr(&self) -> Result<SocketAddr>

Returns the address of this socket's peer.

The connect method will connect the socket to a peer.

fn recv_from(&self, buf: &mut [u8]) -> Result<(usize, SocketAddr)>

Receives data from the socket.

On success, returns the number of bytes read and the address from whence the data came.

fn recv(&self, buf: &mut [u8]) -> Result<usize>

Receives data from the socket.

On success, returns the number of bytes read.

fn send_to<P: AsRef<Path>>(&self, buf: &[u8], path: P) -> Result<usize>

Sends data on the socket to the specified address.

On success, returns the number of bytes written.

fn send(&self, buf: &[u8]) -> Result<usize>

Sends data on the socket to the socket's peer.

The peer address may be set by the connect method, and this method will return an error if the socket has not already been connected.

On success, returns the number of bytes written.

fn shutdown(&self, how: Shutdown) -> Result<()>

Shut down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).