Struct unix_socket::UnixStream

pub struct UnixStream {
    // some fields omitted
}

A stream which communicates over a Unix domain socket.

Methods

impl UnixStream

fn connect<P: AsRef<Path>>(path: P) -> Result<UnixStream>

Connect to the socket named by path.

Linux provides, as a nonportable extension, a separate "abstract" address namespace as opposed to filesystem-based addressing. If path begins with a null byte, it will be interpreted as an "abstract" address. Otherwise, it will be interpreted as a "pathname" address, corresponding to a path on the filesystem.

fn unnamed() -> Result<(UnixStream, UnixStream)>

Create an unnamed pair of connected sockets.

Returns two UnixStreams which are connected to each other.

fn try_clone(&self) -> Result<UnixStream>

Create a new independently owned handle to the underlying socket.

The returned UnixStream is a reference to the same stream that this object references. Both handles will read and write the same stream of data, and options set on one stream will be propogated to the other stream.

fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address of the local half of this connection.

fn peer_addr(&self) -> Result<SocketAddr>

Returns the socket address of the remote half of this connection.

fn shutdown(&self, how: Shutdown) -> Result<()>

Shut down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).