Struct unix_socket::UnixStream

pub struct UnixStream { /* fields omitted */ }

A Unix stream socket.

Examples

use unix_socket::UnixStream;
use std::io::prelude::*;

let mut stream = UnixStream::connect("/path/to/my/socket").unwrap();
stream.write_all(b"hello world").unwrap();
let mut response = String::new();
stream.read_to_string(&mut response).unwrap();
println!("{}", response);Run

Methods

impl UnixStream

fn connect<P: AsRef<Path>>(path: P) -> Result<UnixStream>

Connects to the socket named by path.

Linux provides, as a nonportable extension, a separate "abstract" address namespace as opposed to filesystem-based addressing. If path begins with a null byte, it will be interpreted as an "abstract" address. Otherwise, it will be interpreted as a "pathname" address, corresponding to a path on the filesystem.

fn pair() -> Result<(UnixStream, UnixStream)>

Creates an unnamed pair of connected sockets.

Returns two UnixStreams which are connected to each other.

fn try_clone(&self) -> Result<UnixStream>

Creates a new independently owned handle to the underlying socket.

The returned UnixStream is a reference to the same stream that this object references. Both handles will read and write the same stream of data, and options set on one stream will be propogated to the other stream.

fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address of the local half of this connection.

fn peer_addr(&self) -> Result<SocketAddr>

Returns the socket address of the remote half of this connection.

fn set_read_timeout(&self, timeout: Option<Duration>) -> Result<()>

Sets the read timeout for the socket.

If the provided value is None, then read calls will block indefinitely. It is an error to pass the zero Duration to this method.

fn set_write_timeout(&self, timeout: Option<Duration>) -> Result<()>

Sets the write timeout for the socket.

If the provided value is None, then write calls will block indefinitely. It is an error to pass the zero Duration to this method.

fn read_timeout(&self) -> Result<Option<Duration>>

Returns the read timeout of this socket.

fn write_timeout(&self) -> Result<Option<Duration>>

Returns the write timeout of this socket.

fn set_nonblocking(&self, nonblocking: bool) -> Result<()>

Moves the socket into or out of nonblocking mode.

fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).